backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
google-genai==0.3.0  # Gemini API client library
